Fix stack overflow in Lua
If a text resource is an array, function i18n::_() chooses one element
of the array at random. And then, the chosen element remains in the
stack of Lua until Lua crashes by stack overflow. Since this commit, the
element is popped from the Lua stack after passed to C++ world.