Allow autosave if focus is lost
More will have to be done in order to account for Android's app
lifecycle. Unlike desktop platforms, the application is not
immediately exited when the main SDL thread finishes. The app will
still run in the background, and if it is selected then it will crash
due to attempting to allocate another SDL window (only one is allowed
on Android). For now, if focus is lost and the user is being queried
for an action at the main screen, the game will be quicksaved. (This
is configurable.)