Forcibly exit application on main thread finish
Unlike desktop OSes, Android applications will remain open after
exiting until the OS decides to kill the process to gain resources. A
consequence of this is that any native code ran by SDL has to be
restartable, or the app will crash on startup if it is closed and
reopened. To make the code restartable, all reliance on static
initialization has to be removed, and all variables have to be in a
clean state before initializing again. The specific cause of the crash
is SDL trying to create a second window even though the original is
still active. For some reason (likely the use of static singletons),
the finalizers for SDL structs don't seem to get called when SDL's
thread closes, so those structures stay open.

To work around all that, code is added so the application is forcibly
exited when SDL's thread exits. This is bad because it doesn't obey
Android's process lifecycle, but it will prevent crashing when trying
to reopen the app.

See also: https://forums.libsdl.org/viewtopic.php?p=52445#52445