Create first iteration of dynamic data registry
It uses HCL as a definition format and Lua to store data (as before).
However, it will not be limited to a small set of datatypes, and new
datatypes will be able to be registered by mods. That way other mods
can reuse existing datatypes and add their own seamlessly without the
original mod's knowledge. One example would be registering new
continuous actions: a player can register a new "activity" data
containing important variables and a callback name, which is executed
in Lua.

It is not as fast as having data structures in C++, so for the first
usage of the registry (characters) data is memoized inside an
intermediary database as before, since characters are central to the
game and their instantiation needs to be fast. However, less important
types can be manipulated entirely in Lua.

It currently resolves enum types by querying Lua. The specification of
what values are valid for a datatype haven't been placed in yet, but
it will reuse some of the schema type of the config definition (after
formalizing it more).