Make item.number private and force modifying through getter/setter
Attempting to keep the list of items that are valid in Lua in sync
with the actually valid items in C++ is difficult because item
validity is determined only by whether or not an item's number is 0,
so anything that modifies the item's number can cause its slot to
become unused. To enforce that the lifetime of items is accurate, a
getter and setter are used.