Rethink deferred state setting of removed character
Characters can apparently be used for a good time after their state is
set to empty. This interferes with passing them to Lua as their
references will be removed. In the meantime, defer the state change of
characters killed in damage_hp until the end of the function.

There should be a better solution to this...