Move methods out of sol::usertype initializer to avoid large codegen
Putting too many things into the initializer of a sol::usertype causes
a template instantiation depth error at depth 900.

Problem is, anything declared as a sol variable or property has to be
declared on the usertype at startup, because it overrides __index.
There are still many fields to be added to the LuaCharacter usertype,
so at some point something about this has to change. It may be a
limitation in sol, which would be bad. There has to be some other way
of doing it, because other games exist with even more fields on their
exported objects (like Factorio).