Fix food not getting rotten sometimes if you believe in Kumiromi
Cause:
// pseudocode:
i = (The date and time in-game)
for food in items:
    if (the expiration date) < i:                    (*0)
        The food gets rotten.
        if you believe in Kumiromi:
            i = (The item's subcategory)             (*1)
            if a third:
                You extract seed(s) from the food.
                i = (The date and time in-game)      (*2)

The variable `i` is modified at line (*1) and restored at line (*2), but
(*2) is executed only if you extract seeds. As a result, when a food
gets rotten and you don't extract seed from it, the variable `i` become
an invalid value and then, the condition (*0) fails to determine whether
a food should get rotten or not.