Change handle invalidation to happen right before new ref creation
Before, the lifecycle of Lua handles was as follows:

empty -> empty = no handle
empty -> valid = create handle
valid -> valid = valid handle
valid -> empty = remove handle
empty -> empty = no handle
empty -> valid = create handle

Now, it was changed to this.

empty -> empty = no handle
empty -> valid = create handle
valid -> valid = valid handle
valid -> empty = valid handle
empty -> empty = valid handle
empty -> valid = remove old handle, create new handle

The specific reason for this change was the on_removal callback inside
the new buff datatype. In Death Word, the target character dies before
the on_removal callback runs, causing the passed handle to be nil.
With the new behavior, the handle will still exist after the character
dies, but it will be removed and invalidated when a new character is
created in the same slot. The same applies for items.

The tradeoff of this approach is the need for the user to determine if
the handle is in the "empty" state before operating on it. On the
other hand, it would allow for behaviors like reviving dead
characters.