Change "BreakBeforeBinaryOperators" option to None in .clang-format
There are two allowed values other than "None": "NonAssignment" and
"All". However, the two have a bug, inserting unexpected spaces. It has
been fixed in the latest `clang-format`, but we decide to change the
value itself in order to make both old `clang-format` and new one output
the same result.