Move the position of the crypt of the damned right
The previous position, (37, 20), is not an accessible cell and the
actual position varies in each save data. In most cases, the actual
position become (38, 20) or (37, 21), but sometimes the crypt moves far
away or moves several times when a diastrophism occurs.

The new position, (38, 20), is the same as MMAh's. Thanks to the author
of MMAh and reporters of this bug.