Clean up lua references around item/chara copying
Before, item_copy and chara_copy could copy into slots occupied by Lua
references whose concrete object had already been removed, but which
are still valid. This is due to the changes in e4ef187f causing
Lua references to not be invalidated at the point of removal.

The solution is to check before copying whether the reference in the
item/character slot is valid, and to clear it before copying.

There are probably other bugs of this kind where a stale reference is
not cleared before performing an action that would create a new handle
in the same slot.