Add list support for i18n-checker
a {
    b = [
        "foo",
        "bar",
        "baz",
    ]
}

In this case, call of `i18n::s.get("a.b")` returns one of "foo", "bar",
or "baz" at random. The previous implementation reported "a.b" was
missing because its internal locale table held "a.b.1", "a.b.2", and
"a.b.3".
Now, we check not only `locale_key` but `locale_key .. ".1"`.