Generate Lua API documentation from C++ comments
Running `make docgen` will regenerate most of the API documentation.
This makes it far easier to keep the documentation in sync when the
functions and documentation are kept close together, and helps avoid
missing documentation of new functions. The only documentation not
generated is the EnumKinds page, which ought to be changed to a string
enum eventually.

Rust stable is required to run it, but the generated documentation
will be checked in so it won't be necessary to run it in CI.

Comments with the `@luadoc` will be interpreted as ldoc source
comments and output to doc/api/*.luadoc. The syntax is:

@luadoc NAME DOC_TYPE LUA_FIELD_TYPE

- NAME sets the name of the function. Normally it is taken from the
  parsed function name of the comment.
- DOC_TYPE changes the documentation type. Only used with freestanding
  comments. Only really used value is `field` currently. Used to
  indicate class/module fields. See the lua class sources for
  examples.
- LUA_FIELD_TYPE sets the type of a field. Required if DOC_TYPE is
  `field`.

Functions tagged with `@luadoc` in the lua_api/lua_env directories
will be given function documentation. The name, module and parameters
will be parsed out automatically, so that all that is needed is the
comment.

Also, to document a module, the `@luadoc` comment must be added before
the namespace (LuaApi<...>) in the header of the class/API source. See
the existing sources for examples.

Enums in lua_env/enums/enums.cpp are also documented, through a
different process. Due to a limitation of clang-rs the source file is
tokenized directly and parsed for each enum variant. However, this
avoids the need to type all the variants out manually.