Improve performance of text rendering with shadow
A text with shadow is rendered in this way:

for dy in [-1, 0, 1]:
    for dx in [-1, 0, 1]:
        if dy == 0 and dx == 0:
            continue
        render_text(x + dx, y + dy, text, shadow_color)
render_text(x, dy, text, text_color)

It creates text's texture 5 times in total. The first 4 textures are
exactly the same as each other and must be reusable.
This commit optimizes snail::Renderer::render_text_with_shadow()'s
implementation and let it reuse the same texture.