Fix get_executable_path() on Windows
wcstombs() takes a multibyte character buffer (output), a wide character
buffer, and the length of the output buffer. However, actually, the
length of *wide character buffer* was passed to. Due to that,
get_executable_path() returns a shorter path than the actual one on
Windows. Now, the function does not convert wchar_t* to char* and
returns internal representation on each OS. On Windows, it is UTF-16,
wchar_t*. (Strictly speaking, C++ specification does not say wchar_t* is
encoded in UTF-16 but Windows does so.)