Limit some of rnd()'s range to emulate vanilla rnd() function.
`rnd()` is one of the HSP built-in functions, whose result is capped by
`RAND_MAX` in C lang. To emulate it, some of `rnd()` in foobar's
codebase are substituted by `rnd_capped()` that has the same upper
limit.

`rnd_capped(n)` returns `k` in `[0, m)` where `m` is `min(n, RAND_MAX)`.

The capped version is also added to Lua `Rand` module. The usage is
exactly the same as `Rand.rnd()`.