Disable network feature temporarily until the bug is resolved
Bug:

If you enable network feature on Windows, Elona foobar will freeze. The
whole binary data is surely sent to the remote server, but neither the
success handler nor the failure handler is called. The bug is
reproducible, but the cause hasn't been found yet. Anyway, the network
features are temporarily disabled until the bug resolved.