Remove meaningless statements related to Item::index
In ctrl_file.cpp:
As the member variable `Item::index` is always properly set in the
constructor of `Inventory`, it is unnecessary to restore the member when
loading save files in `ctrl_file.cpp`.

In talk_npc.cpp:
The value of `Item::index` was assigned to the variable `p`, but it was
never used after the assignment.

In command.cpp:
When `Item::number()` reaches 0, `cell_refresh()` is called in
`Item::set_number()`. It is unnecessary to call `cell_refresh()`
manually.