Fix a crash related to item drawing
How to reproduce

1. Put an arbitrary item in a town.
2. Leave there and wait for 5 days.
3. Visit the town.

Description

When a town is refreshed, all items you put on the ground are swept.
However, item display information (`CellData::item_info_memory`) is not
removed. It is an expected behavior, but as a result, causes a crash of
Elona foobar. To draw a removed item, it has to be stored in a save data
somehow. I am working on the issue, but have not completed yet.
As a band-aid, ghost items are not rendered in this version.