Truncate `Item::enchantments`; remove enchantments of ID 0
Before, `Item::enchantments` had 15 elements even if the list had no
enchantments. Enchantment of ID 0 means that the slot is not occupied
and free to use. This commit changes the structure, deleting such slots.

Example:

Before:
[ 10003, 10002, 10001, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0 ]
The length of the list is always 15, and 4th to 15th elements are filled
with 0.

After:
[ 10003, 10002, 10001 ]
The length of the list is variable, and 4th to 15th elements are removed
now.